+++ tools/test.rb
@@ -8,3 +8,5 @@ module Rails
     @root ||= Pathname.new(COMPONENT_ROOT)
   end
 end
+
+Rails::TestUnitReporter.executable = "bin/test"